Creación de estructura del proyecto
Cree la estructura de carpetas que utilizaremos en el proyecto. También movi algunas carpetas para que sepamos donde encontrarlas
