+++ app/src/main/java/com/redb/to_dolist/Vistas/NavigationDrawer/gallery/GalleryFragment.kt
@@ -1,4 +1,4 @@
-package com.redb.to_dolist.ui.tools
+package com.redb.to_dolist.Vistas.NavigationDrawer.gallery
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,20 +10,20 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.redb.to_dolist.R
 
-class ToolsFragment : Fragment() {
+class GalleryFragment : Fragment() {
 
-    private lateinit var toolsViewModel: ToolsViewModel
+    private lateinit var galleryViewModel: GalleryViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        toolsViewModel =
-            ViewModelProviders.of(this).get(ToolsViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_tools, container, false)
-        val textView: TextView = root.findViewById(R.id.text_tools)
-        toolsViewModel.text.observe(this, Observer {
+        galleryViewModel =
+            ViewModelProviders.of(this).get(GalleryViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_gallery, container, false)
+        val textView: TextView = root.findViewById(R.id.text_gallery)
+        galleryViewModel.text.observe(this, Observer {
             textView.text = it
         })
         return root
