+++ app/src/main/java/com/redb/to_dolist/Vistas/NavigationDrawer/gallery/GalleryViewModel.kt
@@ -1,13 +1,13 @@
-package com.redb.to_dolist.ui.home
+package com.redb.to_dolist.Vistas.NavigationDrawer.gallery
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class GalleryViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is gallery Fragment"
     }
     val text: LiveData<String> = _text
 }
