+++ app/src/main/java/com/redb/to_dolist/Vistas/NavigationDrawer/home/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package com.redb.to_dolist.ui.share
+package com.redb.to_dolist.Vistas.NavigationDrawer.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ShareViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is share Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
