+++ app/src/main/java/com/redb/to_dolist/Vistas/NavigationDrawer/send/SendFragment.kt
@@ -1,4 +1,4 @@
-package com.redb.to_dolist.ui.slideshow
+package com.redb.to_dolist.Vistas.NavigationDrawer.send
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,20 +10,20 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.redb.to_dolist.R
 
-class SlideshowFragment : Fragment() {
+class SendFragment : Fragment() {
 
-    private lateinit var slideshowViewModel: SlideshowViewModel
+    private lateinit var sendViewModel: SendViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        slideshowViewModel =
-            ViewModelProviders.of(this).get(SlideshowViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_slideshow, container, false)
-        val textView: TextView = root.findViewById(R.id.text_slideshow)
-        slideshowViewModel.text.observe(this, Observer {
+        sendViewModel =
+            ViewModelProviders.of(this).get(SendViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_send, container, false)
+        val textView: TextView = root.findViewById(R.id.text_send)
+        sendViewModel.text.observe(this, Observer {
             textView.text = it
         })
         return root
