+++ app/src/main/java/com/redb/to_dolist/Vistas/NavigationDrawer/share/ShareViewModel.kt
@@ -1,13 +1,13 @@
-package com.redb.to_dolist.ui.slideshow
+package com.redb.to_dolist.Vistas.NavigationDrawer.share
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class ShareViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is share Fragment"
     }
     val text: LiveData<String> = _text
 }
