+++ app/src/main/java/com/redb/to_dolist/Vistas/NavigationDrawer/slideshow/SlideshowViewModel.kt
@@ -1,13 +1,13 @@
-package com.redb.to_dolist.ui.gallery
+package com.redb.to_dolist.Vistas.NavigationDrawer.slideshow
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class GalleryViewModel : ViewModel() {
+class SlideshowViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is gallery Fragment"
+        value = "This is slideshow Fragment"
     }
     val text: LiveData<String> = _text
 }
