+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,37 +7,37 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.redb.to_dolist.ui.home.HomeFragment"
+        android:name="com.redb.to_dolist.Vistas.NavigationDrawer.home.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="com.redb.to_dolist.ui.gallery.GalleryFragment"
+        android:name="com.redb.to_dolist.Vistas.NavigationDrawer.gallery.GalleryFragment"
         android:label="@string/menu_gallery"
         tools:layout="@layout/fragment_gallery" />
 
     <fragment
         android:id="@+id/nav_slideshow"
-        android:name="com.redb.to_dolist.ui.slideshow.SlideshowFragment"
+        android:name="com.redb.to_dolist.Vistas.NavigationDrawer.slideshow.SlideshowFragment"
         android:label="@string/menu_slideshow"
         tools:layout="@layout/fragment_slideshow" />
 
     <fragment
         android:id="@+id/nav_tools"
-        android:name="com.redb.to_dolist.ui.tools.ToolsFragment"
+        android:name="com.redb.to_dolist.Vistas.NavigationDrawer.tools.ToolsFragment"
         android:label="@string/menu_tools"
         tools:layout="@layout/fragment_tools" />
 
     <fragment
         android:id="@+id/nav_share"
-        android:name="com.redb.to_dolist.ui.share.ShareFragment"
+        android:name="com.redb.to_dolist.Vistas.NavigationDrawer.share.ShareFragment"
         android:label="@string/menu_share"
         tools:layout="@layout/fragment_share" />
 
     <fragment
         android:id="@+id/nav_send"
-        android:name="com.redb.to_dolist.ui.send.SendFragment"
+        android:name="com.redb.to_dolist.Vistas.NavigationDrawer.send.SendFragment"
         android:label="@string/menu_send"
         tools:layout="@layout/fragment_send" />
 </navigation>
